the second parameter to fivebeans.runner can now be an object or a pathname
This path lets you pass an object to fivebeans.runner instead of the pathname of the config file.  The big advantage of this is that it lets you set the options dynamically: set the tubes to watch based on the environment, populate the handlers list with a fs.readdirSync, et cetera.